Add unit files and install scripts for code-server
These provide a useful setup, through a myriad of proxies. After setup
of `tsproxy@code-server`, going to `https://code-server.<tailnet>.ts.net`
causes a systemd socket activation for
`.local/tsproxy/code-server.sock`.

That starts a `systemd-socket-proxyd` unit that in turn starts
`code-server` and connects to it. The socket activation also starts a watchdog
timer (another systemd unit). When the watchdog expires, the associated
service checks if the  `code-server` heartbeat file
has been updated recently; if so, the script reschedules the watchdog.
If the heartbeat file is stale, there are no recent connections to
`code-server`, so the script shuts down the unit.